tests: move methods from repository to a manager
