+++ src/App/Handler/User/SetUserPasswordHandler.php
@@ -6,9 +6,9 @@ use App\Command\User\SetUserPasswordCommand;
 use App\Event\CommandEvent;
 use App\Event\NotificationEvent;
 use App\Handler\BaseDoctrineHandler;
+use App\Service\User\UserManager;
 use Doctrine\Common\Persistence\ManagerRegistry;
 use JMS\DiExtraBundle\Annotation as DI;
-use Salt\UserBundle\Entity\User;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 
@@ -19,9 +19,15 @@ use Symfony\Component\Validator\Validator\ValidatorInterface;
  */
 class SetUserPasswordHandler extends BaseDoctrineHandler
 {
-    public function __construct(ValidatorInterface $validator, ManagerRegistry $registry)
+    /**
+     * @var UserManager
+     */
+    private $userManager;
+
+    public function __construct(ValidatorInterface $validator, ManagerRegistry $registry, UserManager $userManager)
     {
         parent::__construct($validator, $registry);
+        $this->userManager = $userManager;
     }
 
     /**
@@ -42,7 +48,7 @@ class SetUserPasswordHandler extends BaseDoctrineHandler
         $username = $command->getUserName();
         $plainPassword = $command->getPlainPassword();
 
-        $newPassword = $this->em->getRepository(User::class)->setUserPassword($username, $plainPassword);
+        $newPassword = $this->userManager->setUserPassword($username, $plainPassword);
 
         $command->setPlainPassword($newPassword);
