+++ tests/_support/Helper/UserManagement.php
@@ -2,6 +2,7 @@
 
 namespace Helper;
 
+use App\Service\User\UserManager;
 use Codeception\Module\Symfony;
 use Codeception\TestInterface;
 use Doctrine\ORM\EntityManager;
@@ -94,6 +95,8 @@ class UserManagement extends \Codeception\Module
         /** @var EntityManager $em */
         $em = $symfony->grabService('doctrine.orm.default_entity_manager');
 
+        $userManager = $symfony->grabService(UserManager::class);
+
         /** @var \Faker\Generator $faker */
         $faker = \Faker\Factory::create();
 
@@ -114,7 +117,7 @@ class UserManagement extends \Codeception\Module
 
         if ($user) {
             $username = $user->getUsername();
-            $userRepo->setUserPassword($username, $password);
+            $userManager->setUserPassword($username, $password);
             $em->flush();
         } else {
             $orgRepo = $em->getRepository(Organization::class);
@@ -132,7 +135,7 @@ class UserManagement extends \Codeception\Module
             }
 
             $username = 'TEST:'.$role.':'.$faker->userName;
-            $userRepo->addNewUser($username, $org, $password, $role, $status);
+            $userManager->addNewUser($username, $org, $password, $role, $status);
             $em->flush();
 
             $user = $userRepo->createQueryBuilder('u')
